+++ src/common/lib/client/realtime.ts
@@ -60,8 +60,8 @@ class Channels extends EventEmitter {
   constructor(realtime: Realtime) {
     super();
     this.realtime = realtime;
-    this.all = {};
-    this.inProgress = {};
+    this.all = Object.create(null);
+    this.inProgress = Object.create(null);
     realtime.connection.connectionManager.on('transport.active', () => {
       this.onTransportActive();
     });